updated os_core.cpp

+++ CSOPESY_Group7_MCO1/os_core.cpp
@@ -934,7 +934,7 @@ void ConsoleManager::schedulerFCFS() {
         if (schedulerPaused)
             continue;
 
-        for (int i = 0; i < cpuCores.size(); ++i) {
+        for (size_t i = 0; i < cpuCores.size(); ++i) {
             if (!cpuCores[i] && !waitingQueue.empty()) {
                 process_console* nextProcess = waitingQueue.front();
                 waitingQueue.pop();
@@ -971,7 +971,7 @@ void ConsoleManager::schedulerRR() {
         if (schedulerPaused)
             continue;
 
-        for (int i = 0; i < cpuCores.size(); ++i) {
+        for (size_t i = 0; i < cpuCores.size(); ++i) {
             if (!cpuCores[i] && !waitingQueue.empty()) {
                 process_console* nextProcess = waitingQueue.front();
                 waitingQueue.pop();
@@ -1003,4 +1003,5 @@ void ConsoleManager::schedulerRR() {
         }
     }
 }
+
